relocate files api
