+++ packages/main-api/src/controllers/v3/spaces/files/update_files.rs
@@ -2,14 +2,13 @@ use crate::controllers::v3::spaces::dto::*;
 use crate::models::space::SpaceCommon;
 
 use crate::models::user::User;
-use crate::types::{EntityType, Partition, TeamGroupPermission};
+use crate::types::{Partition, TeamGroupPermission};
 use crate::{AppState, Error2};
 
 use aide::NoApi;
 
-use crate::models::file::SpaceFile;
+use crate::features::spaces::files::SpaceFile;
 use crate::types::File;
-use crate::types::space_file_feature_type::SpaceFileFeatureType;
 use axum::extract::{Json, Path, State};
 use bdk::prelude::*;
 
@@ -17,8 +16,6 @@ use bdk::prelude::*;
 pub struct UpdateSpaceFileRequest {
     #[schemars(description = "Space Files")]
     pub files: Vec<File>,
-    #[schemars(description = "Space File Feature Type: (Overview, Recommendation)")]
-    pub feature_type: SpaceFileFeatureType,
 }
 
 #[derive(Debug, Clone, Default, serde::Serialize, serde::Deserialize, JsonSchema)]
@@ -48,34 +45,22 @@ pub async fn update_files_handler(
         return Err(Error2::NoPermission);
     }
 
-    let files = SpaceFile::get(
-        &dynamo.client,
-        &space_pk.clone(),
-        Some(EntityType::SpaceFile(req.feature_type.to_string())),
-    )
-    .await?;
+    let (pk, sk) = SpaceFile::keys(&space_pk);
+
+    let files = SpaceFile::get(&dynamo.client, &pk.clone(), Some(sk.clone())).await?;
 
     if files.is_some() {
-        SpaceFile::updater(
-            &space_pk.clone(),
-            EntityType::SpaceFile(req.feature_type.to_string()),
-        )
-        .with_feature_type(req.feature_type.clone())
-        .with_files(req.files)
-        .execute(&dynamo.client)
-        .await?;
+        SpaceFile::updater(&pk.clone(), sk.clone())
+            .with_files(req.files)
+            .execute(&dynamo.client)
+            .await?;
     } else {
-        let files = SpaceFile::new(space_pk.clone(), req.feature_type.clone(), req.files);
+        let files = SpaceFile::new(space_pk.clone(), req.files);
 
         files.create(&dynamo.client).await?;
     }
 
-    let files = SpaceFile::get(
-        &dynamo.client,
-        &space_pk.clone(),
-        Some(EntityType::SpaceFile(req.feature_type.to_string())),
-    )
-    .await?;
+    let files = SpaceFile::get(&dynamo.client, &pk.clone(), Some(sk.clone())).await?;
 
     let files = files.unwrap();
