+++ packages/main-api/src/features/spaces/files/mod.rs
@@ -1,0 +1,2 @@
+pub mod models;
+pub use models::*;
