+++ packages/main-api/src/features/spaces/files/models/mod.rs
@@ -1,3 +1,2 @@
 pub mod space_file;
-
 pub use space_file::*;
