+++ packages/main-api/src/features/spaces/mod.rs
@@ -1,3 +1,4 @@
 pub mod discussions;
+pub mod files;
 pub mod polls;
 pub mod recommendations;
