+++ packages/main-api/src/models/dynamo_tables/main/space/mod.rs
@@ -1,5 +1,4 @@
 pub mod deliberation;
-pub mod file;
 pub mod space_common;
 
 pub use deliberation::*;
